requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
instagrapi==1.16.27
Pillow==9.2.0
pycodestyle==2.9.1
pycryptodomex==3.15.0
pydantic==1.9.1
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
